+++ Apple-Frameworks/Apple-Frameworks/Screens/FrameworkDetailView/FrameworkDetailView.swift
@@ -16,17 +16,7 @@ struct FrameworkDetailView: View {
   var body: some View {
     
     VStack {
-      HStack {
-        Spacer()
-        Button {
-          isShowingDetailView = false
-        } label: {
-          Image(systemName: "xmark")
-            .foregroundColor(Color(.label))
-            .imageScale(.large)
-            .frame(width: 44, height: 44)
-        }
-      }.padding()
+      XDismissButton(isShowingModal: $isShowingDetailView)
       
       Spacer()
