+++ Apple-Frameworks/Apple-Frameworks/Screens/FrameworkGridView/FrameworkGridView.swift
@@ -11,15 +11,11 @@ struct FrameworkGridView: View {
   
   @StateObject var viewModel = FrameworkGridViewModel()
   
-  let columns: [GridItem] = [GridItem(.flexible()),
-                             GridItem(.flexible()),
-                             GridItem(.flexible())]
-  
   var body: some View {
     
     NavigationView {
       ScrollView {
-        LazyVGrid(columns: columns) {
+        LazyVGrid(columns: viewModel.columns) {
           ForEach(MockData.frameworks) { framework in
             FrameworkTitleView(framework: framework)
               .onTapGesture {
@@ -42,25 +38,5 @@ struct FrameworkGridView_Previews: PreviewProvider {
   static var previews: some View {
     FrameworkGridView()
       .preferredColorScheme(.dark)
-  }
-}
-
-struct FrameworkTitleView: View {
-  
-  let framework: Framework
-  
-  var body: some View {
-    VStack {
-      Image(framework.imageName)
-        .resizable()
-        .aspectRatio(contentMode: .fit)
-        .frame(width: 90, height: 90)
-      Text(framework.name)
-        .font(.title2)
-        .fontWeight(.semibold)
-        .scaledToFit()
-        .minimumScaleFactor(0.5)
-    }
-    .padding()
   }
 }
