+++ Apple-Frameworks/Apple-Frameworks/Screens/FrameworkGridView/FrameworkGridViewModel.swift
@@ -10,10 +10,12 @@ import SwiftUI
 final class FrameworkGridViewModel: ObservableObject {
   
   var selectedFramework: Framework? {
-    didSet {
-      isShowingDetailView = true
-    }
+    didSet { isShowingDetailView = true }
   }
   
   @Published var isShowingDetailView = false
+  
+  let columns: [GridItem] = [GridItem(.flexible()),
+                             GridItem(.flexible()),
+                             GridItem(.flexible())]
 }
