+++ Apple-Frameworks/Apple-Frameworks/Views/View/FrameworkTitleView.swift
@@ -1,0 +1,36 @@
+//
+//  FrameworkTitleView.swift
+//  Apple-Frameworks
+//
+//  Created by Dan Smith on 08/11/2020.
+//
+
+import SwiftUI
+
+struct FrameworkTitleView: View {
+  
+  let framework: Framework
+  
+  var body: some View {
+    VStack {
+      Image(framework.imageName)
+        .resizable()
+        .aspectRatio(contentMode: .fit)
+        .frame(width: 90, height: 90)
+      Text(framework.name)
+        .font(.title2)
+        .fontWeight(.semibold)
+        .scaledToFit()
+        .minimumScaleFactor(0.5)
+    }
+    .padding()
+  }
+}
+
+struct FrameworkTitleView_Previews: PreviewProvider {
+    static var previews: some View {
+      FrameworkTitleView(framework: MockData.sampleFramework)
+        .padding()
+        .previewLayout(.sizeThatFits)
+    }
+}
